+++ backend/src/controllers/meta.controller.ts
@@ -1,15 +1,35 @@
 import { NextFunction, Request, Response } from "express";
 import metaService from "../services/meta.service";
 import { checkValidation } from "../middleware/validation.middleware";
-import { MetaResponseDTO, MetaUpdateDTO } from "../dtos/meta.dto";
+import { MaintenanceResponseDTO, MaintenanceUpdateDTO, MetaResponseDTO, MetaUpdateDTO } from "../dtos/meta.dto";
 import { IRepoError } from "../repositories/interfaces/error.repository.interface";
 
 class MetaController {
-    async getMetaData(req: Request, res: Response, next: NextFunction) {
+    async getMetaById(req: Request, res: Response, next: NextFunction) {
         try {
             checkValidation(req);
             const id: number = +(req.params.id);
-            const response: MetaResponseDTO | IRepoError | null = await metaService.getMetaData(id);
+            const response: MetaResponseDTO | IRepoError | null = await metaService.getMetaById(id);
+            res.send(response);
+        } catch(err: any) {
+            next(err);
+        }
+    }
+
+    async getMetaByName(req: Request, res: Response, next: NextFunction) {
+        try {
+            checkValidation(req);
+            const name: string = req.params.name;
+            const response: MetaResponseDTO | IRepoError | null = await metaService.getMetaByName(name);
+            res.send(response);
+        } catch(err: any) {
+            next(err);
+        }
+    }
+
+    async getAllData(req: Request, res: Response, next: NextFunction) {
+        try {
+            const response: MetaResponseDTO[] | IRepoError | null = await metaService.getAllData();
             res.send(response);
         } catch(err: any) {
             next(err);
@@ -22,6 +42,29 @@ class MetaController {
             const id: number = +(req.params.id);
             const dto: MetaUpdateDTO = req.body;
             const response: MetaResponseDTO | IRepoError | null = await metaService.updateMetaData(id, dto);
+            res.send(response);
+        } catch(err: any) {
+            next(err);
+        }
+    }
+
+    async getMaintenanceMode(req: Request, res: Response, next: NextFunction) {
+        try {
+            checkValidation(req);
+            const name: string = req.params.name;
+            const response: MaintenanceResponseDTO | IRepoError | null = await metaService.getMaintenanceMode(name);
+            res.send(response);
+        } catch(err: any) {
+            next(err);
+        }
+    }
+
+    async setMaintenanceMode(req: Request, res: Response, next: NextFunction) {
+        try {
+            checkValidation(req);
+            const name: string = req.params.name;
+            const dto: MaintenanceUpdateDTO = req.body;
+            const response: MaintenanceResponseDTO | IRepoError | null = await metaService.setMaintenanceMode(name, dto);
             res.send(response);
         } catch(err: any) {
             next(err);